created sub dirs for mvvm model and moved view files to view dir, converters to converter dir
