+++ source/DotNetBay.WPF/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application x:Class="DotNetBay.WPF.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="MainWindow.xaml">
+             StartupUri="View/MainWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
