+++ source/DotNetBay.WPF/Converter/DebugConverter.cs
@@ -1,12 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Data;
 
-namespace DotNetBay.WPF
+namespace DotNetBay.WPF.Converter
 {
     public class DebugConverter : IValueConverter
     {
