+++ source/DotNetBay.WPF/View/BidView.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="DotNetBay.WPF.BidView"
+﻿<Window x:Class="DotNetBay.WPF.View.BidView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:wpftk="http://schemas.xceed.com/wpf/xaml/toolkit"
