+++ source/DotNetBay.WPF/View/SellView.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="DotNetBay.WPF.SellView"
+﻿<Window x:Class="DotNetBay.WPF.View.SellView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:wpftk="http://schemas.xceed.com/wpf/xaml/toolkit"
